pyproject.toml
[tool.poetry]
name = "library-management-web-application"
version = "0.1.0"
description = "A library management application made using Django."
authors = ["Antony Nyagah <tony.m.nyagah@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "library_management_web_application" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
whitenoise = "^6.5.0"
gunicorn = "^21.2.0"
pygraphviz = "^1.11"
pyparsing = "^3.1.0"
pydot = "^1.4.2"
pytest = "^7.4.1"
pytest-django = "^4.5.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
commitizen = "^3.2.2"
pre-commit = "^3.3.2"
django-extensions = "^3.2.3"

[tool.ruff]
exclude = ["vendor/*", "*/migrations/*"]
line-length = 130

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "library_project.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
